# dist/openapi.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: Localhost (dev)
    url: http://localhost:3000
info:
  description: This is an API designed for users to create, share and explore recipes and create family recipes
  version: 1.0.2
  title: Recipe Sharing API
  contact:
    name: Abed
    email: showganabed@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /recipes/search:
    get:
      tags:
        - recipes
      summary: Search for recipes with optional filters
      operationId: searchRecipes
      description: |
        Search recipes by name and optionally filter by cuisine, diet, intolerance, and results number, latest search results are displayed if they are available.
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Recipe name or keywords to search for
        - name: cuisine
          in: query
          required: false
          schema:
            type: string
          description: Type of cuisine to filter by (e.g., Italian, Mexican)
        - name: diet
          in: query
          required: false
          schema:
            type: string
          description: Type of diet to filter by (e.g., vegetarian, keto)
        - name: intolerances
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated list of intolerances (e.g., gluten, dairy)
        - name: number
          in: query
          required: false
          schema:
            type: integer
            enum: [5, 10, 15]
            default: 5
          description: Number of recipes to return (default is 5)
      responses:
        '200':
          description: List of recipes matching the search criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
        '400':
          description: Invalid search parameters
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User registered
        '400':
          description: Validation error
  /auth/login:
    post:
      tags: [auth]
      summary: Log in a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid credentials
  /auth/logout:
    post:
      tags: [auth]
      summary: Log out a user
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully
...
  /recipes/create:
    post:
      tags:
        - recipes
      summary: Create a new recipe (custom format)
      operationId: createRecipe
      description: |
        Creates a new recipe and stores it in the internal database.
        Fields must match what the backend expects including extendedIngredients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomNewRecipe'
      responses:
        '201':
          description: Recipe successfully created
        '400':
          description: Invalid input or missing required fields
  /recipes/{recipeId}:
    get:
      tags:
        - recipes
      summary: Get full recipe details
      operationId: getRecipeById
      description: |
        Returns the full recipe details including preview fields, ingredients list,
        instructions, and number of meals.
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: string
          description: ID of the recipe to retrieve
      responses:
        '200':
          description: Full recipe details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          description: Recipe not found

    delete:
      tags:
        - recipes
      summary: Delete a recipe by ID
      operationId: deleteRecipe
      parameters:
        - name: recipeId
          in: path
          required: true
          description: ID of the recipe to delete from the database
          schema:
            type: integer
      responses:
        '200':
          description: Recipe deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Recipe not found or unauthorized
components:
  securitySchemes: 
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - username
        - firstname
        - lastname
        - country
        - password
        - password_confirmation
        - email
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 8
          pattern: '^[A-Za-z]+$'
        firstname:
          type: string
        lastname:
          type: string
        country:
          type: string
          description: Use values from https://restcountries.com/v3.1/all
        password:
          type: string
          pattern: '^(?=.*[0-9])(?=.*[^A-Za-z0-9]).{5,10}$'
        password_confirmation:
          type: string
        email:
          type: string
          format: email
    UserLogin:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    NewRecipe:
      type: object
      required:
        - recipe_image
        - recipe_title
        - prep_duration
        - popularity
        - vegetarian
        - vegan
        - gluten_free
        - ingredients
        - instructions
        - amount_of_meals
      properties:
        recipe_image:
          type: string
          format: uri
          example: https://example.com/image.jpg
        recipe_title:
          type: string
          example: Chocolate Cake
        prep_duration:
          type: string
          example: 00:45:00
          description: Preparation time in HH:MM:SS format
        popularity:
          type: integer
          example: 152
        vegetarian:
          type: boolean
        vegan:
          type: boolean
        gluten_free:
          type: boolean
        clicked_by_user:
          type: boolean
        saved_by_user:
          type: boolean
        ingredients:
          type: array
          description: List of ingredients and their amounts
          items:
            type: object
            required: [name, amount]
            properties:
              name:
                type: string
                example: Sugar
              amount:
                type: string
                example: "2 cups"
        instructions:
          type: string
          example: |
            1. Preheat the oven to 180°C.
            2. Mix all ingredients.
            3. Bake for 30 minutes.
        amount_of_meals:
          type: integer
          example: 4
    Recipe:
      allOf:
        - $ref: '#/components/schemas/NewRecipe'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              example: 123e4567-e89b-12d3-a456-426614174000
...
    RecipePreview:
      type: object
      required:
        - recipe_image
        - recipe_title
        - prep_duration
        - popularity
        - vegetarian
        - vegan
        - gluten_free
        - clicked_by_user
        - saved_by_user
      properties:
        recipe_image:
          type: string
          format: uri
        recipe_title:
          type: string
        prep_duration:
          type: string
          example: "00:30:00"
        popularity:
          type: integer
        vegetarian:
          type: boolean
        vegan:
          type: boolean
        gluten_free:
          type: boolean
        clicked_by_user:
          type: boolean
        saved_by_user:
          type: boolean
    CustomNewRecipe:
      type: object
      required:
        - recipe_title
        - recipe_image
        - prep_duration
        - vegetarian
        - vegan
        - gluten_free
        - amount_of_meals
        - instructions
        - extendedIngredients
      properties:
        recipe_title:
          type: string
          example: My New Recipe
        recipe_image:
          type: string
          format: uri
          example: https://example.com/image.jpg
        prep_duration:
          type: string
          example: 30
        vegetarian:
          type: boolean
        vegan:
          type: boolean
        gluten_free:
          type: boolean
        amount_of_meals:
          type: integer
          example: 4
        instructions:
          type: string
          example: |
            1. Chop onions
            2. Fry with oil
            3. Add spices
            4. Serve
        extendedIngredients:
          type: array
          description: List of ingredients and their amounts
          items:
            type: object
            required: [ name, amount ]
            properties:
              name:
                type: string
                example: Onion
              amount:
                type: string
                example: 1 large